minor: Get helpers working with e2e tests
