+++ lib/helpers/os.js
@@ -63,11 +63,10 @@ const assignIn = require('lodash/assignIn');
 const mapValues = require('lodash/mapValues');
 
 const Bluebird = require('bluebird');
-const config = require('../config');
+const { config } = require('../config');
 const imagefs = require('balena-image-fs');
 const { fs } = require('mz');
 const { join } = require('path');
-const tmp = require('tmp');
 const pipeline = Bluebird.promisify(require('stream').pipeline);
 const zlib = require('zlib');
 const util = require('util');
@@ -76,13 +75,13 @@ const util = require('util');
 const injectBalenaConfiguration = async (image, configuration, partition = 1) => {
 	await imagefs.interact(image, partition, async (_fs) => {
 		return util.promisify(_fs.writeFile)(
-				'/config.json',
-				JSON.stringify(configuration),
-				{ flag: 'w' }
-			)
-		}).catch((err) => {
-				return undefined;
-		});
+			'/config.json',
+			JSON.stringify(configuration),
+			{ flag: 'w' }
+		)
+	}).catch((err) => {
+		return undefined;
+	});
 };
 
 // TODO: This function should be implemented using Reconfix
@@ -111,8 +110,8 @@ const injectNetworkConfiguration = async (image, configuration, partition = 1) =
 		'method=auto',
 	];
 
-	if(configuration.wireless.interfaceName){
-		wifiConfiguration.splice(2,0,`${configuration.wireless.interfaceName}`,)
+	if (configuration.wireless.interfaceName) {
+		wifiConfiguration.splice(2, 0, `${configuration.wireless.interfaceName}`,)
 	}
 
 	if (configuration.wireless.psk) {
@@ -129,12 +128,12 @@ const injectNetworkConfiguration = async (image, configuration, partition = 1) =
 
 	await imagefs.interact(image, partition, async (_fs) => {
 		return util.promisify(_fs.writeFile)(
-				'/system-connections/balena-wifi',
-				wifiConfiguration.join('\n'),
-				{ flag: 'w' }
-			)
-		}).catch((err) => {
-			return undefined;
+			'/system-connections/balena-wifi',
+			wifiConfiguration.join('\n'),
+			{ flag: 'w' }
+		)
+	}).catch((err) => {
+		return undefined;
 	});
 };
