+++ lib/index.ts
@@ -1,8 +1,8 @@
-import { Worker } from './helpers/worker';
+import { Worker } from './helpers/worker.js';
 import { BalenaOS } from './helpers/os.js';
 import { Sdk } from './helpers/sdk';
 import { preload } from './helpers/preload';
-import { Utils } from './helpers/utils';
+import { utils } from './helpers/utils';
 import { add, getStream } from './helpers/archiver';
 import { blockhash, hammingDistance } from './helpers/screen-capture';
 
@@ -10,7 +10,7 @@ export {
     Worker, 
     BalenaOS, 
     Sdk, 
-    Utils,
+    utils,
     preload,
     add, getStream,
     blockhash,
